Move _utils directory from src/app/ to src/
