+++ src/app/episodes/page/[page]/page.tsx
@@ -1,7 +1,7 @@
 import '../../../layout.css'
 import EpisodeEntry from '../../../../components/EpisodeEntry';
 import { Episode } from '../../../../components/types/Episode';
-import { FeedLoader } from '../../../_utils/FeedLoader';
+import { FeedLoader } from '../../../../utils/FeedLoader';
 import Link from 'next/link';
 import React from 'react';
